Added update and delete function

+++ src/main/java/com/example/eCom/Exception/ProductNotFound.java
@@ -1,7 +1,7 @@
 package com.example.eCom.Exception;
 
 public class ProductNotFound extends RuntimeException{
-    public ProductNotFound(int id){
+    public ProductNotFound(long id){
         super(String.format("Product not found with the given id %d",id ));
     }
 }
